requirements.txt
streamlit
scikit-learn
pandas
numpy
matplotlib
wordcloud
joblib
newspaper3k
youtube-transcript-api
instaloader

